backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.7.4
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
groq==0.4.1
openai==1.12.0
requests==2.31.0

